Support simple redirect page after login

+++ html/code/modules/authsystem/xaradmin/modifyconfig.php
@@ -24,12 +24,12 @@ function authsystem_admin_modifyconfig()
     // Security
     if (!xarSecurityCheck('AdminAuthsystem')) return;
     
-    if (!xarVarFetch('phase',        'str:1:100', $phase,       'modify', XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
-    if (!xarVarFetch('uselockout',   'checkbox',  $data['uselockout'],  xarModVars::get('authsystem', 'uselockout'),     XARVAR_NOT_REQUIRED)) return;
-    if (!xarVarFetch('lockouttime',  'int:1:',    $data['lockouttime'], (int)xarModVars::get('authsystem', 'lockouttime'),       XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
-    if (!xarVarFetch('lockouttries', 'int:1:',    $data['lockouttries'], (int)xarModVars::get('authsystem', 'lockouttries'),       XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
-    if (!xarVarFetch('forwarding_page', 'str',    $data['forwarding_page'], (int)xarModVars::get('authsystem', 'forwarding_page'),       XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
-    if (!xarVarFetch('ask_forward',     'checkbox',  $data['ask_forward'],  xarModVars::get('authsystem', 'ask_forward'),     XARVAR_NOT_REQUIRED)) return;
+    if (!xarVarFetch('phase',           'str:1:100', $phase,                  'modify', XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
+    if (!xarVarFetch('uselockout',      'checkbox',  $data['uselockout'],      xarModVars::get('authsystem', 'uselockout'),     XARVAR_NOT_REQUIRED)) return;
+    if (!xarVarFetch('lockouttime',     'int:1:',    $data['lockouttime'],     (int)xarModVars::get('authsystem', 'lockouttime'),       XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
+    if (!xarVarFetch('lockouttries',    'int:1:',    $data['lockouttries'],    (int)xarModVars::get('authsystem', 'lockouttries'),       XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
+    if (!xarVarFetch('forwarding_page', 'str',       $data['forwarding_page'], xarModVars::get('authsystem', 'forwarding_page'),       XARVAR_NOT_REQUIRED, XARVAR_PREP_FOR_DISPLAY)) return;
+    if (!xarVarFetch('ask_forward',     'checkbox',  $data['ask_forward'],     xarModVars::get('authsystem', 'ask_forward'),     XARVAR_NOT_REQUIRED)) return;
 
     $data['module_settings'] = xarMod::apiFunc('base','admin','getmodulesettings',array('module' => 'authsystem'));
     $data['module_settings']->setFieldList('items_per_page, use_module_alias, module_alias_name, enable_short_urls, frontend_page');
